+++ MADDecoder.h
@@ -50,7 +50,7 @@
 
 - (SeekIndex *)seekIndex;
 - (int)getNextOverlayedBeepSampleAtTime:(double)time;
-- (void)writePCMData:(void *)dataPtr length:(unsigned long)length;
+- (void)writePCMData:(void *)dataPtr length:(size_t)length;
 - (BOOL)canContinueDecoding;
 - (void)foundSilenceFrom:(double)start to:(double)end;
 - (void)decodingErrorOverflow;
